API-227766 Replace deprecated integration API calls with new Modules API API-227766-144450 [DEV] Replace deprecated integration API calls with new Modules API
