+++ onevizion/__init__.py
@@ -64,8 +64,8 @@ from onevizion.export import Export
 
 from onevizion.EMail import EMail
 
-from onevizion.integration.log import IntegrationLog
-from onevizion.integration.loglevel import LogLevel
+from onevizion.module.log import IntegrationLog
+from onevizion.module.loglevel import LogLevel
 
 if sys.version_info.major >= 3 and sys.version_info.minor >= 4:
 	from onevizion.notif.service import NotificationService
