--- onevizion/integration/__init__.py
@@ -1,2 +1,0 @@
-from onevizion.integration import loglevel
-from onevizion.integration import log
