+++ onevizion/module/__init__.py
@@ -1,0 +1,2 @@
+from onevizion.module import loglevel
+from onevizion.module import log
